fix(traders-orb): Int8 overflow in time arithmetic broke ALL ORB detection
ROOT CAUSE for TRADERS_ORB placing zero orders since deployment:

  pl.col('time').dt.hour()  returns Int8 (max=127)
  Int8 * 60 overflows for any hour >= 3 (3*60=180 > 127)
  -> filter expression silently evaluates to false
  -> range_bars query returns 0 rows
  -> 'too few range bars' check triggers, processing aborts

Proof: bars_at_hour_13 == 60, but (hour*60+min)<15 returned 0 rows.
For hour 13, 13*60 = 780 which wraps in Int8 arithmetic.

Fix: cast(pl.Int32) before multiplying, matching the pattern already used
in DataEnricher.enrich_with_opening_range (src/application/calculators.py
line 113). Two occurrences in _process_orb in traders_live_engine.py.

This bug explains why ORB never traded despite multiple symbols breaking
out today (USOUSD broke +$2.85 above r_high, UKOUSD +$2.66, etc.).

+++ src/execution/traders_live_engine.py
@@ -449,8 +449,8 @@ class TradersEngine:
             range_end_min = oh * 60 + om + rm
             range_bars = m1.filter(
                 (pl.col("time").dt.date() == m1.tail(1)["time"][0].date())
-                & ((pl.col("time").dt.hour() * 60 + pl.col("time").dt.minute()) >= (oh * 60 + om))
-                & ((pl.col("time").dt.hour() * 60 + pl.col("time").dt.minute()) < range_end_min)
+                & ((pl.col("time").dt.hour().cast(pl.Int32) * 60 + pl.col("time").dt.minute()) >= (oh * 60 + om))
+                & ((pl.col("time").dt.hour().cast(pl.Int32) * 60 + pl.col("time").dt.minute()) < range_end_min)
             )
             if len(range_bars) < max(1, rm // 2):
                 continue
